requirements.txt
annotated-types==0.7.0
anyio==4.4.0
attrs==24.1.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==43.0.0
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.15.4
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
idna==3.7
iniconfig==2.0.0
isodate==0.6.1
lxml==5.2.2
more-itertools==10.3.0
numpy==2.0.1
openai==1.38.0
packaging==24.1
platformdirs==4.2.2
pluggy==1.5.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
PyJWT==2.9.0
pytest==8.3.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
safetensors==0.4.3
simple-salesforce==1.12.6
six==1.16.0
sniffio==1.3.1
tokenizers==0.19.1
tomli==2.0.1
tqdm==4.66.5
transformers==4.43.3
typing_extensions==4.12.2
urllib3==2.2.2
zeep==4.2.1
